fix: complete image autoresizing
The problem where memegenerator panel was
getting squeezed too small because of
'margin: 0 auto' has also been fixed
by addition of an outer container.
Note sure why that fixes the issue but
probabaly marghin: 0 auto was interfering
with the fact that it was a flex item so
align-items: stretch (the default) was
not working properly.